tenant_subfolder/requirements.txt
amqp==5.0.9
APScheduler==3.8.1
arrow==1.2.1
asgiref==3.4.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
blessed==1.19.1
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.0
Django==4.0
django-apscheduler==0.6.0
django-celery-results==2.2.0
django-cors-headers==3.10.1
django-picklefield==3.0.1
django-tenants==3.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
fatoora==2.1.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.3
MarkupSafe==2.0.1
packaging==21.3
Pillow==9.0.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
pydantic==1.8.2
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyzbar==0.1.8
qrcode==7.3.1
redis==3.5.3
requests==2.26.0
ruamel.yaml==0.17.19
ruamel.yaml.clib==0.2.6
schedule==1.1.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.7
uttlv==0.6.0
validators==0.18.2
vine==5.0.0
wcwidth==0.2.5
